+++ tests/dish_vcc_configuration/test_load_dish_cfg.py
@@ -4,18 +4,14 @@ import pytest
 from pytest_bdd import given, scenario, then, when
 from tango import DevState
 
+from tests.resources.test_harness.constant import COMMAND_COMPLETED
 from tests.resources.test_harness.helpers import (
     device_attribute_changed,
     get_master_device_simulators,
     prepare_json_args_for_centralnode_commands,
 )
-from tests.resources.test_support.common_utils.result_code import ResultCode
 
 
-@pytest.mark.skip(
-    reason="Test fails because of TRANSIENT CORBA EXCEPTION with command"
-    + " exceeding the timeout of 3 seconds"
-)
 @pytest.mark.SKA_mid
 @scenario(
     "../features/load_dish_cfg_command.feature",
@@ -25,7 +21,8 @@ def test_dish_id_vcc_configuration():
     """This test validate that TMC is able to load the dish vcc
     configuration file provided to LoadDishCfg command.
     Validate that k-numbers set on dish masters
-    Validate sysParam and sourceSysParam attribute set on csp master leaf node
+    Validate dishVccConfig and sourceDishVccConfig attribute set on csp master
+    leaf node
     """
 
 
@@ -87,70 +84,75 @@ def invoke_load_dish_cfg(
     assert event_recorder.has_change_event_occurred(
         central_node_mid.central_node,
         "longRunningCommandResult",
-        (unique_id[0], str(int(ResultCode.OK))),
+        (unique_id[0], COMMAND_COMPLETED),
         lookahead=5,
     )
 
 
 @then("TMC should pass the configuration to CSP Controller")
 def tmc_pass_configuration_to_csp_controller(simulator_factory):
-    """Validate sysParam and sourceSysParam attribute set on Csp Master
+    """Validate dishVccConfig and sourceDishVccConfig attribute set on Csp
+    Master
     :param simulator_factory: fixture for creating simulator devices for
     mid Telescope respectively.
     """
-    csp_master_sim, _, _, _, _ = get_master_device_simulators(
+    csp_master_sim, _, _, _, _, _ = get_master_device_simulators(
         simulator_factory
     )
-    expected_sys_param = {
-        "interface": "https://schema.skao.int"
-        "/ska-mid-cbf-initial-parameters/2.2",
+    expected_dish_vcc_config = {
+        "interface": "https://schema.skao.int/ska-mid-cbf-initsysparam/1.0",
         "dish_parameters": {
-            "SKA001": {"vcc": 1, "k": 11},
-            "SKA036": {"vcc": 2, "k": 101},
+            "SKA001": {"k": 119, "vcc": 1},
+            "SKA036": {"k": 1127, "vcc": 2},
+            "SKA063": {"k": 620, "vcc": 3},
+            "SKA100": {"k": 101, "vcc": 4},
         },
     }
-    assert json.loads(csp_master_sim.sysParam) == expected_sys_param
+    assert json.loads(csp_master_sim.dishVccConfig) == expected_dish_vcc_config
 
 
 @then("TMC displays the current version of Dish and VCC configuration")
-def validate_sys_param_attribute_set(central_node_mid):
-    """Valdate sysParam and sourceSysParam attribute
+def validate_dish_vcc_config_attribute_set(central_node_mid):
+    """Valdate dishVccConfig and sourceDishVccConfig attribute
     correctly set on csp master leaf node
     :param central_node_mid: fixture for a TMC CentralNode Mid under test
     which provides simulated master devices
     """
-    interface_schema = (
-        "https://schema.skao.int/ska-mid-cbf-initial-parameters/2.2"
-    )
-    expected_sys_param = json.dumps(
+    interface_schema = "https://schema.skao.int/ska-mid-cbf-initsysparam/1.0"
+    expected_dish_vcc_config = json.dumps(
         {
             "interface": interface_schema,
             "dish_parameters": {
-                "SKA001": {"vcc": 1, "k": 11},
-                "SKA036": {"vcc": 2, "k": 101},
+                "SKA001": {"k": 119, "vcc": 1},
+                "SKA036": {"k": 1127, "vcc": 2},
+                "SKA063": {"k": 620, "vcc": 3},
+                "SKA100": {"k": 101, "vcc": 4},
             },
         }
     )
-    expected_source_sys_param = json.dumps(
+    expected_source_dish_vcc_config = json.dumps(
         {
             "interface": interface_schema,
             "tm_data_sources": [
-                "gitlab://gitlab.com/ska-telescope/ska-telmodel-data?main#"
-                + "tmdata"
+                "car://gitlab.com/ska-telescope/ska-telmodel-data?ska-sdp-"
+                + "tmlite-repository-1.0.0#tmdata"
             ],
             "tm_data_filepath": (
-                "instrument/dishid_vcc_configuration/mid_cbf_parameters.json"
+                "instrument/ska1_mid_psi/ska-mid-cbf-system-parameters.json"
             ),
         }
     )
 
     assert device_attribute_changed(
         device=central_node_mid.csp_master_leaf_node,
-        attribute_name_list=["sysParam", "sourceSysParam"],
-        attribute_value_list=[expected_sys_param, expected_source_sys_param],
+        attribute_name_list=["dishVccConfig", "sourceDishVccConfig"],
+        attribute_value_list=[
+            expected_dish_vcc_config,
+            expected_source_dish_vcc_config,
+        ],
         timeout=100,
     ), (
-        "sysParam and sourceSysParam attribute value is not set "
+        "dishVccConfig and sourceDishVccConfig attribute value is not set "
         "on csp master leaf node"
     )
 
@@ -167,6 +169,7 @@ def validate_k_number_set(simulator_factory):
         dish_master_1_sim,
         dish_master_2_sim,
         _,
+        _,
     ) = get_master_device_simulators(simulator_factory)
-    assert dish_master_1_sim.kValue == 11
-    assert dish_master_2_sim.kValue == 101
+    assert dish_master_1_sim.kValue == 119
+    assert dish_master_2_sim.kValue == 1127