+++ tests/dish_vcc_configuration/test_load_dish_cfg_negative_scenario.py
@@ -2,6 +2,8 @@
 Test case to validate negative scenario for
    Dish Vcc map configuration feature
 """
+import json
+
 import pytest
 from pytest_bdd import given, parsers, scenario, then, when
 from tango import DevState
@@ -20,10 +22,6 @@ from tests.resources.test_support.common_utils.result_code import ResultCode
 # @pytest.mark.skip("Need design change to avoid timeout corba exception")
 
 
-@pytest.mark.skip(
-    reason="Test fails because of TRANSIENT CORBA EXCEPTION with command"
-    + " exceeding the timeout of 3 seconds"
-)
 @pytest.mark.SKA_mid
 @scenario(
     "../features/load_dish_cfg_command_negative_scenario.feature",
@@ -44,10 +42,6 @@ def test_central_node_return_error_for_invalid_file():
     """
 
 
-@pytest.mark.skip(
-    reason="Test fails because of TRANSIENT CORBA EXCEPTION with command"
-    + " exceeding the timeout of 3 seconds"
-)
 @pytest.mark.SKA_mid
 @scenario(
     "../features/load_dish_cfg_command_negative_scenario.feature",
@@ -59,10 +53,6 @@ def test_central_node_return_error_for_invalid_dish_id():
     """
 
 
-@pytest.mark.skip(
-    reason="Test fails because of TRANSIENT CORBA EXCEPTION with command"
-    + " exceeding the timeout of 3 seconds"
-)
 @pytest.mark.SKA_mid
 @scenario(
     "../features/load_dish_cfg_command_negative_scenario.feature",
@@ -74,10 +64,6 @@ def test_central_node_return_error_for_duplicate_vcc_id():
     """
 
 
-@pytest.mark.skip(
-    reason="Test fails because of TRANSIENT CORBA EXCEPTION with command"
-    + " exceeding the timeout of 3 seconds"
-)
 @pytest.mark.SKA_mid
 @pytest.mark.skip(
     reason="This will be enable once Dish Vcc feature is integrate."
@@ -89,8 +75,8 @@ def test_central_node_return_error_for_duplicate_vcc_id():
 def test_central_node_handle_exception():
     """This test validate that when exception is raised on csp controller
     device then the csp master should raise the error to central node and
-    sysParam and sourceSysParam attributes are not updated at csp master leaf
-    node device
+    dishVccConfig and sourceDishVccConfig attributes are not updated at
+    csp master leaf node device
     """
 
 
@@ -140,6 +126,30 @@ def test_tmc_rejects_command_with_error(error_message):
     """Test validate that command failed with error message"""
     assert pytest.command_result_code == ResultCode.REJECTED
     assert error_message in pytest.command_result_message[0]
+
+
+@then(
+    parsers.parse(
+        "TMC updates longrunningcommandresult with error {error_message}"
+    )
+)
+def test_validates_longrunningcommandresult_with_error(
+    error_message, central_node_mid, event_recorder
+):
+    """Test validate that command failed with error message"""
+    # Subscribe for longRunningCommandResult attribute
+    event_recorder.subscribe_event(
+        central_node_mid.central_node, "longRunningCommandResult"
+    )
+    assert event_recorder.has_change_event_occurred(
+        central_node_mid.central_node,
+        "longRunningCommandResult",
+        (
+            pytest.command_result_message[0],
+            json.dumps([ResultCode.FAILED, error_message]),
+        ),
+        lookahead=5,
+    )
 
 
 @then(
@@ -229,9 +239,11 @@ def invoke_command_load_cfg_on_defective_csp(
     csp_sim = simulator_factory.get_or_create_simulator_device(
         SimulatorDeviceType.MID_CSP_MASTER_DEVICE
     )
-    pytest.initial_sysParam = central_node_mid.csp_master_leaf_node.sysParam
-    pytest.initial_sourceSysParam = (
-        central_node_mid.csp_master_leaf_node.sourceSysParam
+    pytest.initial_dishVccConfig = (
+        central_node_mid.csp_master_leaf_node.dishVccConfig
+    )
+    pytest.initial_sourceDishVccConfig = (
+        central_node_mid.csp_master_leaf_node.sourceDishVccConfig
     )
 
     csp_sim.SetDefective(ERROR_PROPAGATION_DEFECT)
@@ -255,20 +267,20 @@ def invoke_command_load_cfg_on_defective_csp(
 
 
 @then(
-    "sysParam and sourceSysParam attributes "
+    "dishVccConfig and sourceDishVccConfig attributes "
     "remains unchanged on CSP Master Leaf Node"
 )
 def check_sys_param_source_sys_param_attributes(central_node_mid):
-    """Test validate that sysParam and sourceSysParam attributes
+    """Test validate that dishVccConfig and sourceDishVccConfig attributes
     are not updated after error
     """
     assert (
-        pytest.initial_sysParam
-        == central_node_mid.csp_master_leaf_node.sysParam
+        pytest.initial_dishVccConfig
+        == central_node_mid.csp_master_leaf_node.dishVccConfig
     )
     assert (
-        pytest.initial_sourceSysParam
-        == central_node_mid.csp_master_leaf_node.sourceSysParam
+        pytest.initial_sourceDishVccConfig
+        == central_node_mid.csp_master_leaf_node.sourceDishVccConfig
     )