+++ tests/resources/test_harness/utils/common_utils.py
@@ -1,12 +1,26 @@
 """This module implement common utils
 """
+
+from __future__ import annotations
+
 import time
 from os.path import dirname, join
+from typing import TYPE_CHECKING
 
+from assertpy import assert_that
 from ska_control_model import ObsState
+from ska_tango_testing.integration import TangoEventTracer
+from tango import DevState
 
 from tests.resources.test_harness.constant import COMMAND_COMPLETED
+from tests.resources.test_harness.utils.constant import EVENT_TIMEOUT
 from tests.resources.test_harness.utils.wait_helpers import Waiter
+from tests.resources.test_support.enum import DishMode
+
+if TYPE_CHECKING:
+    from tests.resources.test_harness.central_node_mid import (
+        CentralNodeWrapperMid,
+    )
 
 
 def get_subarray_input_json(slug):
@@ -339,3 +353,116 @@ def check_configure_successful_csp(
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_node, "obsState", ObsState.READY, lookahead=10
     )
+
+
+def setup_dish_events(
+    central_node_mid,
+    event_tracer,
+    dish_ids,
+) -> None:
+    """
+    This function will subscribe events for dish attributes
+
+    Args:
+        central_node_mid: Fixture for a TMC CentralNode wrapper class
+        event_tracer: Fixture for EventTracer class
+    """
+    dish_ids = ["SKA001", "SKA036", "SKA063", "SKA100"]
+    events = ["dishMode", "pointingState"]
+
+    for dish_id in dish_ids:
+        dish_master = central_node_mid.dish_master_dict[dish_id]
+        dish_leaf = central_node_mid.dish_leaf_node_dict[dish_id]
+
+        for event in events:
+            event_tracer.subscribe_event(dish_master, event)
+            event_tracer.subscribe_event(dish_leaf, event)
+
+
+def turn_on_telescope(
+    central_node_mid: CentralNodeWrapperMid,
+    event_tracer: TangoEventTracer,
+    dish_ids: str,
+) -> None:
+    """
+    A method to put Telescope ON
+
+    Args:
+        central_node_mid: Fixture for a TMC CentralNode wrapper class
+        event_tracer: Fixture for EventTracer class
+        dish_ids: Dish Ids to be assigned to Subarray
+    """
+
+    assert central_node_mid.csp_master.ping() > 0
+    assert central_node_mid.sdp_master.ping() > 0
+    for dish_id in dish_ids.split(","):
+        assert central_node_mid.dish_master_dict[dish_id].ping() > 0
+        assert central_node_mid.dish_leaf_node_dict[dish_id].ping() > 0
+
+    setup_dish_events(central_node_mid, event_tracer, dish_ids)
+    event_tracer.subscribe_event(central_node_mid.csp_master, "State")
+    event_tracer.subscribe_event(central_node_mid.sdp_master, "State")
+    event_tracer.subscribe_event(
+        central_node_mid.central_node, "telescopeState"
+    )
+
+    central_node_mid.move_to_on()
+
+    assert_that(event_tracer).described_as(
+        "FAILED ASSUMPTION AFTER ON COMMAND: "
+        "CSP Master device"
+        f"({central_node_mid.csp_master.dev_name()}) "
+        "is expected to be in DevState ON",
+    ).within_timeout(EVENT_TIMEOUT).has_change_event_occurred(
+        central_node_mid.csp_master,
+        "State",
+        DevState.ON,
+    )
+
+    assert_that(event_tracer).described_as(
+        "FAILED ASSUMPTION AFTER ON COMMAND: "
+        "SDP Master device"
+        f"({central_node_mid.sdp_master.dev_name()}) "
+        "is expected to be in DevState ON",
+    ).within_timeout(EVENT_TIMEOUT).has_change_event_occurred(
+        central_node_mid.sdp_master,
+        "State",
+        DevState.ON,
+    )
+
+    for dish_id in ["SKA001", "SKA036", "SKA063", "SKA100"]:
+
+        assert_that(event_tracer).described_as(
+            "FAILED ASSUMPTION AFTER ON COMMAND: "
+            "Dish Master device"
+            f"({central_node_mid.dish_master_dict[dish_id].dev_name()}) "
+            "is expected to be in DishMode STANDBY_FP",
+        ).within_timeout(EVENT_TIMEOUT).has_change_event_occurred(
+            central_node_mid.dish_master_dict[dish_id],
+            "dishMode",
+            DishMode.STANDBY_FP,
+        )
+
+        assert_that(event_tracer).described_as(
+            "FAILED ASSUMPTION AFTER ON COMMAND: "
+            "Dish Leaf Node device"
+            f"({central_node_mid.dish_leaf_node_dict[dish_id].dev_name()}) "
+            "is expected to be in DishMode STANDBY_FP",
+        ).within_timeout(EVENT_TIMEOUT).has_change_event_occurred(
+            central_node_mid.dish_leaf_node_dict[dish_id],
+            "dishMode",
+            DishMode.STANDBY_FP,
+        )
+
+    assert_that(event_tracer).described_as(
+        "FAILED ASSUMPTION AFTER ON COMMAND: "
+        "Central Node "
+        f"({central_node_mid.central_node.dev_name()}) "
+        "is expected to be in telescopeState ON",
+    ).within_timeout(EVENT_TIMEOUT).has_change_event_occurred(
+        central_node_mid.central_node,
+        "telescopeState",
+        DevState.ON,
+    )
+
+    event_tracer.clear_events()