+++ tests/tmc_csp_new_ITH/test_telescope_state.py
@@ -35,7 +35,6 @@ def test_on_to_off():
     """Test transitioning from ON to OFF."""
 
 
-@pytest.mark.skip(reason="CBF is not supporting STANDBY command")
 @pytest.mark.tmc_csp_new_ITH
 @scenario(
     "../tmc_csp_new_ITH/features/telescope_state.feature",
@@ -138,9 +137,9 @@ def verify_off_state(
     tmc: TMCFacade,
     csp: CSPFacade,
 ):
-    """The telescope and CSP devices transition to the OFF state."""
+    """TMC and CSP devices transition to the OFF state."""
     assert_that(event_tracer).described_as(
-        "The telescope and CSP devices should transition from ON to OFF state."
+        "TMC and CSP devices should transition from ON to OFF state."
     ).within_timeout(ASSERTIONS_TIMEOUT).has_change_event_occurred(
         tmc.central_node,
         "telescopeState",
@@ -158,18 +157,12 @@ def verify_standby_state(
     tmc: TMCFacade,
     csp: CSPFacade,
 ):
-    """The telescope and CSP devices transition to the STANDBY state."""
-    assert_that(event_tracer).described_as(
-        "The telescope and CSP master should transition "
-        "to the STANDBY state. "
-        "CSP subarray should transition to OFF state."
-    ).within_timeout(ASSERTIONS_TIMEOUT).has_change_event_occurred(
+    """TMC should transition to the STANDBY state."""
+    assert_that(event_tracer).within_timeout(ASSERTIONS_TIMEOUT).described_as(
+        "TMC should transition to the STANDBY state. "
+    ).has_change_event_occurred(
         tmc.central_node,
         "telescopeState",
-        DevState.STANDBY,
-    ).has_change_event_occurred(
-        csp.csp_master,
-        "State",
         DevState.STANDBY,
     )
 
@@ -180,9 +173,9 @@ def verify_on_state(
     tmc: TMCFacade,
     csp: CSPFacade,
 ):
-    """The telescope and CSP devices transition to the ON state."""
+    """TMC and CSP devices transition to the ON state."""
     assert_that(event_tracer).described_as(
-        "The telescope and CSP devices should transition " "to the ON state."
+        "TMC and CSP devices should transition to the ON state."
     ).within_timeout(ASSERTIONS_TIMEOUT).has_change_event_occurred(
         tmc.central_node,
         "telescopeState",