+++ CHANGELOG.md
@@ -3,6 +3,16 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 Added
 -----
+[0.22.8-rc.1]
+*************
+* Updated ska-tmc-centralnode v0.16.8 to resolve skb-495
+* Updated ska-tmc-cspleafnode v0.21.2 to resolve skb-495 and skb-530 
+
+[0.22.7-rc.1]
+*************
+* Updated ska-tmc-dishleafnode v0.17.6 to resolve SKB-511 and SKB-467
+* Track command will not be invoked from Dish leaf node if pointingState of Dish is TRACK/SLEW
+
 [0.22.6-rc.1]
 *************
 * Updated ska-tmc-subarraynode v.0.23.4 to resolve SKB-509
@@ -11,7 +21,7 @@ Added
 [0.22.5]
 **********
 * Updated dish leaf node version 0.17.3 to resolve SKB-502
-* Updated central node to 0.16.4 to resolve SKB-434
+* Updated central node to 0.16.4 to resolve SKB-434   
 
 [0.22.4]
 ************